The Regen command now removes the sql entries from the database. A little more efficient.

+++ src/main/java/me/hasenzahn1/structurereloot/database/tables/BlockTable.java
@@ -3,6 +3,7 @@ package me.hasenzahn1.structurereloot.database.tables;
 import me.hasenzahn1.structurereloot.database.LootBlockValue;
 import me.hasenzahn1.structurereloot.databasesystem.Database;
 import me.hasenzahn1.structurereloot.databasesystem.Table;
+import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.NamespacedKey;
 import org.bukkit.World;
@@ -12,14 +13,21 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.stream.Collectors;
 
 public class BlockTable extends Table {
 
     private final World world;
 
+    private boolean cacheRemove;
+    private final ArrayList<LootBlockValue> cachedLootBlockValues;
+
     public BlockTable(Database database, World world) {
         super("blocks", database);
         this.world = world;
+
+        cacheRemove = false;
+        cachedLootBlockValues = new ArrayList<>();
     }
 
     @Override
@@ -33,6 +41,10 @@ public class BlockTable extends Table {
     }
 
     public void removeLootBlockValue(LootBlockValue value){
+        if(cacheRemove){
+            cachedLootBlockValues.add(value);
+            return;
+        }
         Connection con = getConnection();
         try(PreparedStatement statement = con.prepareStatement(
                 "DELETE FROM " + getTableName() + " WHERE location='" + value.getLocationString() + "'"
@@ -100,6 +112,32 @@ public class BlockTable extends Table {
         return new ArrayList<>();
     }
 
+    private void removeCachedLbvs(){
+        if(cachedLootBlockValues.size() == 0) return;
+        Connection con = getConnection();
+        StringBuilder sqlString = new StringBuilder("DELETE FROM ").append(getTableName()).append(" WHERE location IN ('").append(cachedLootBlockValues.get(0).getLocationString()).append("'");
 
+        for(int i = 1; i < cachedLootBlockValues.size(); i++){
+            sqlString.append(", '").append(cachedLootBlockValues.get(i).getLocationString()).append("'");
+        }
+        sqlString.append(")");
+        Bukkit.broadcastMessage(sqlString.toString());
+        try(PreparedStatement statement = con.prepareStatement(
+                sqlString.toString()
+        )){
+            statement.execute();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public void setCacheRemove(boolean value){
+        this.cacheRemove = value;
+        if(!value) {
+            Bukkit.broadcastMessage(cachedLootBlockValues.stream().map(LootBlockValue::getLocationString).collect(Collectors.toList()) + "");
+            removeCachedLbvs();
+            cachedLootBlockValues.clear();
+        }
+    }
 
 }
